+++ src/client/cl_main.c
@@ -3233,7 +3233,7 @@ serverStatus_t *CL_GetServerStatus(netadr_t from)
 * @param[in] maxLen
 * @return
 */
-int CL_ServerStatus(const char *serverAddress, char *serverStatusString, size_t maxLen)
+int CL_ServerStatus(const char *serverAddress, char *serverStatusString, unsigned int maxLen)
 {
 	netadr_t       to;
 	serverStatus_t *serverStatus;
@@ -3583,7 +3583,7 @@ void CL_GlobalServers_f(void)
  * @param[in] buflen
  * @param[out] pingtime
  */
-void CL_GetPing(int n, char *buf, size_t buflen, int *pingtime)
+void CL_GetPing(int n, char *buf, unsigned int buflen, int *pingtime)
 {
 	const char *str;
 	int        time;
@@ -3629,7 +3629,7 @@ void CL_GetPing(int n, char *buf, size_t buflen, int *pingtime)
  * @param[out] buf
  * @param[in] buflen
  */
-void CL_GetPingInfo(int n, char *buf, size_t buflen)
+void CL_GetPingInfo(int n, char *buf, unsigned int buflen)
 {
 	if (n < 0 || n >= MAX_PINGREQUESTS || !cl_pinglist[n].adr.port)
 	{
