+++ src/game/g_main.c
@@ -5185,7 +5185,7 @@ void G_MapVoteInfoWrite()
  * @param[out] s
  * @param[in] size
  */
-void G_ReadConfigFileString(char **cnf, char *s, size_t size)
+void G_ReadConfigFileString(char **cnf, char *s, unsigned int size)
 {
 	char *t;
